+++ src/tests/integration/index.ts
@@ -1,0 +1,23 @@
+import * as path from "path";
+import { glob } from "glob";
+import Mocha from "mocha";
+
+export function run(): Promise<void> {
+  const mocha = new Mocha({
+    ui: "tdd",
+    color: true
+  });
+
+  const testsRoot = path.resolve(__dirname);
+
+  return new Promise(async (resolve, reject) => {
+    const files = await glob("**/*.test.js", { cwd: testsRoot });
+    files.forEach((file) => mocha.addFile(path.resolve(testsRoot, file)));
+
+    try {
+      mocha.run((failures) => (failures > 0 ? reject(new Error(`${failures} tests failed.`)) : resolve()));
+    } catch (err) {
+      reject(err);
+    }
+  });
+}